requirements.txt
django==4.2
djangorestframework==3.15.2
django-filter==23.5
psycopg2-binary==2.9.9
python-decouple==3.8
requests==2.31.0
django-redis==5.4.0